% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_utils.R
\name{calculate_bias}
\alias{calculate_bias}
\title{Calculate bias weights}
\usage{
calculate_bias(form, dat, phi.control, thr.rel, time, site_id, ptype, ...)
}
\arguments{
\item{form}{a formula describing the learning task}

\item{dat}{A data frame with time and site_id columns.}

\item{phi.control}{list of parameters as returned
by function \code{UBL::phi.control}}

\item{thr.rel}{a relevance threshold above which an 
observation is considered relevant.}

\item{time}{the column name of the time-stamp}

\item{site_id}{the name of the column containing location IDs}

\item{ptype}{Type of spatio-temporal bias to calculate.
Can include "orig", "addPhiLook", "addNoPhi"}

\item{...}{Parameters to pass to 
\code{sample_wts}. Can include weight type (defaults to orig), 
alpha (defaults to 0.5), and location info (sites_sf or lat+lon+crs).}
}
\value{
Vector of bias weights
}
\description{
Calculate bias weights
}
