% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_utils.R
\name{calculate_bumps}
\alias{calculate_bumps}
\title{Calculate bumps}
\usage{
calculate_bumps(y.relev, thr.rel)
}
\arguments{
\item{y.relev}{Vector containing relevance of sorted target values}

\item{thr.rel}{A number indicating the relevance threshold below which a 
case is considered as belonging to the normal "class".}
}
\value{
Vector containing the first index of each bump
}
\description{
Calculate bumps
}
