% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_utils.R
\name{calculate_obs_ind}
\alias{calculate_obs_ind}
\title{Calculate list of indices in each bump}
\usage{
calculate_obs_ind(s.y, bumps)
}
\arguments{
\item{s.y}{Vector of sorted target values}

\item{bumps}{Vector containing first index of each bump}
}
\value{
List where each entry has a named vector of target values
belonging to a single bump
}
\description{
Calculate list of indices in each bump
}
