% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_utils.R
\name{check_cperc}
\alias{check_cperc}
\title{Check and re-calculate C.perc if necessary}
\usage{
check_cperc(C.perc, inds, type)
}
\arguments{
\item{C.perc}{Value or vector of values with re-sampling percentages.}

\item{inds}{A list with entries obs.ind (a list with named vector cointaining
target values for each bump), und, and ove (containing the indices of 
obs.ind that correspond to bumps of extreme and normal values, respectively).}

\item{type}{"under", "over" or "gauss"}
}
\value{
A vector of re-sampling percentages for each bump that will be 
subject to re-sampling
}
\description{
Check and re-calculate C.perc if necessary
}
