% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_utils.R
\name{gauss.exs}
\alias{gauss.exs}
\title{Provide examples with added Gaussian noise}
\usage{
gauss.exs(exs, bump.exs, pert, exclude = integer(0))
}
\arguments{
\item{exs}{Data frame with observations to perturb}

\item{bump.exs}{Data frame with observations to use as baseline sd}

\item{pert}{Percentage of the sd of bump.exs that should be used as
sd for the gaussian noise added to exs}

\item{exclude}{Indices of columns to exclude from Gaussian noise 
(e.g., spatio-temporal coordinates)}
}
\value{
Data frame where all numeric values have added Gaussian noise
(except for the columns in exclude)
}
\description{
Provide examples with added Gaussian noise
}
