% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_utils.R
\name{get_best}
\alias{get_best}
\title{Get index of best results}
\usage{
get_best(grid.res, metrics, metrics.max, stat)
}
\arguments{
\item{grid.res}{data frame with parameters and metrics results}

\item{metrics}{vector of (up to 2) metrics that should be 
used to decide best internal results}

\item{metrics.max}{vector of Boolean value indicating whether metrics 
should be maximized for best results (in corresponding order to metrics)}

\item{stat}{Stat that should be used to decide on best result
(can be "MED" (median), "MEAN", "IQR" or "SD")}
}
\value{
the row index of the best results
}
\description{
Get index of best results
}
