% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_utils.R
\name{get_inds}
\alias{get_inds}
\title{Calculate indices of relevance bumps}
\usage{
get_inds(s.y, y.relev, pc, thr.rel)
}
\arguments{
\item{s.y}{Vector of sorted target values}

\item{y.relev}{Vector containing relevance of sorted target values}

\item{pc}{Relevance phi.control}

\item{thr.rel}{A number indicating the relevance threshold below which a 
case is considered as belonging to the normal "class".}
}
\value{
A list with entries obs.ind (a list with named vector cointaining
target values for each bump), und, and ove (containing the indices of 
obs.ind that correspond to bumps of extreme and normal values, respectively)
}
\description{
Calculate indices of relevance bumps
}
