% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_weight.R
\name{get_phi_time_wts}
\alias{get_phi_time_wts}
\title{Calculate temporally-biased re-sampling weights with phi look-up}
\usage{
get_phi_time_wts(df, phi, rel.thr, time, site_id)
}
\arguments{
\item{df}{a data frame}

\item{phi}{a vector of the relevance values of 
\code{df}'s target variable}

\item{rel.thr}{a relevance threshold above which an 
observation is considered relevant}

\item{time}{the column name of the time-stamp}

\item{site_id}{the name of the column containing location IDs}
}
\value{
A vector of temporally-biased re-sampling weights, scaled
to fit within range [0,1].
}
\description{
Calculate weights for re-sampling with a temporal bias.
As time goes by past an extreme value, the weight tends to 1.
While no extreme value appears, the weight stays at 0 
(meaning they are less likely to be kept).
More distant to previous rare value: w -> 1; 
very close to recent rare value / no rare values so far: w -> 0.
}
\author{
Mariana Oliveira
}
