% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_utils.R
\name{ratio2cperc}
\alias{ratio2cperc}
\title{Calculate C.perc from ratio}
\usage{
ratio2cperc(ratio, y.relev, rel.thr, type)
}
\arguments{
\item{ratio}{Value that defines the target ratio between the number of 
extreme values and the number of normal values. If more than one bump exists:
re-sampling will be done proportionally to each bump, so that the ratio between
the sum of all extreme values and all normal values in the final data set
is the target ratio.}

\item{y.relev}{A vector containing the relevance values of target}

\item{rel.thr}{A relevance threshold}

\item{type}{"under", "over", or "gauss"}
}
\value{
A list with C.perc, the re-sampling percentage that will result 
in the target ratio and orig_ratio indicating the original ratio between
extreme and normal values
}
\description{
Calculate C.perc from ratio
}
