% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_weight.R
\name{sample_wts}
\alias{sample_wts}
\title{Get spatio-temporal re-sampling weights}
\usage{
sample_wts(
  form,
  df,
  phi.control,
  site_id,
  time,
  rel.thr,
  ptype,
  alpha,
  sites_sf = NULL,
  lon = NULL,
  lat = NULL,
  crs = NULL
)
}
\arguments{
\item{form}{a formula describing the learning task}

\item{df}{a data frame}

\item{phi.control}{list of parameters as returned
by function \code{UBL::phi.control}}

\item{site_id}{the name of the column containing location IDs}

\item{time}{the column name of the time-stamp}

\item{rel.thr}{a relevance threshold above which an 
observation is considered relevant}

\item{ptype}{type of spatio-temporal bias to calculate}

\item{alpha}{weighting parameter for temporal and spatial
re-sampling probabilities. Default 0.5}

\item{sites_sf}{An sf obejct containing station and IDs and 
geometry points of the locations. As an alternative, provide
\code{lon}, \code{lat}, and \code{crs}}

\item{lon}{the name of the column containing the location's longitude}

\item{lat}{the name of the column containing the location's latitude}

\item{crs}{the code for the Coordinate Reference System}
}
\value{
a data.frame with relevance \code{phi},
temporally biased weights \code{time_wts},
and spatially biased weights \code{space_wts} for
each row in \code{df}.
}
\description{
A function that calculates different weights for
re-sampling that is temporally and/or spatially biased.
}
\details{
\code{phi} gives the target variable's relevance 
(higher relevance: phi -> 1; lower relevance: phi -> 0);
\code{time_wts} gives the observation's temporally biased
re-sampling weight (most recent observations: w -> 1; 
oldest: w -> 0.); \code{space_wts} gives the observation's
spatially biased re-sampling weight (farthest away from other 
relevant cases at time slice: d -> 1.).
High \code{time_wts} or \code{space_wts} means the observation is
more likely to be kept.
}
\seealso{
\code{\link{get_phi}}, \code{\link{get_time_wts}},
 \code{\link{get_phi_space_wts}}.
}
