% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_utils.R
\name{select_cases}
\alias{select_cases}
\title{Select cases for adding / keeping in data set}
\usage{
select_cases(obs.ind, C.perc, repl, prob = NULL, epsilon = 1e-04)
}
\arguments{
\item{obs.ind}{Named vector of target values}

\item{C.perc}{Percentage of cases to select from obs.ind}

\item{repl}{A Boolean value controlling whether replication is allowed
when re-sampling observations. Defaults to FALSE when under-sampling
and to TRUE when over-sampling or adding Gaussian noise.}

\item{prob}{Named vector of selection probability. Defaults to NULL.}

\item{epsilon}{Mimimum probability value to add to prob after normalization.}
}
\value{
Vector containing selected observation names
}
\description{
Select cases for adding / keeping in data set
}
